Update README.md
Flask Chat App is a real-time chat application built using Flask and WebSockets. It enables users to engage in instant messaging within a chatroom-based system. The application is designed to be lightweight, fast, and easy to deploy. Flask handles the backend, while WebSockets provide seamless, real-time communication between users.

This project is perfect for those looking to understand how to implement real-time features in a web application using Flask and WebSockets.